Extract AppState and EventLoopRefs into src/app.rs
- Move AppState struct and all its impl methods to dedicated app module
- Move EventLoopRefs struct to app module
- main.rs reduced from ~400+ LOC to 285 LOC
- app.rs is 372 LOC — both well under 500 LOC threshold
- Update imports in main.rs to use app module
